# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Forma API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/runs/{run_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}": {
      "get": {
        "summary": "Project Workspace",
        "operationId": "project_workspace_api_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Artifact": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "component_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "kind": {
            "enum": [
              "step",
              "glb",
              "plot"
            ],
            "title": "Kind",
            "type": "string"
          },
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "storage_path": {
            "title": "Storage Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "revision_id",
          "component_id",
          "name",
          "kind",
          "bytes",
          "storage_path"
        ],
        "title": "Artifact",
        "type": "object"
      },
...
      "CalculationRecord": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          },
          "result": {
            "$ref": "#/components/schemas/CalculationResult"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "reproducible": {
            "title": "Reproducible",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "run_id",
          "revision_id",
          "result",
          "stale",
          "reproducible",
          "created_at"
        ],
        "title": "CalculationRecord",
        "type": "object"
      },
...
      "ChatMessage": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "run_id",
          "role",
          "content",
          "created_at"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Manifest": {
        "additionalProperties": false,
        "properties": {
          "schemaVersion": {
            "const": 1,
            "default": 1,
            "title": "Schemaversion",
            "type": "integer"
          },
          "units": {
            "const": "mm",
            "default": "mm",
            "title": "Units",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "maxItems": 200,
            "title": "Components",
            "type": "array"
          },
          "instances": {
            "items": {
              "$ref": "#/components/schemas/Instance"
            },
            "maxItems": 1000,
            "title": "Instances",
            "type": "array"
          },
          "rootComponentId": {
            "anyOf": [
              {
                "pattern": "^[a-zA-Z][a-zA-Z0-9_-]{0,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Rootcomponentid"
          }
        },
        "required": [
          "schemaVersion",
          "units",
          "components",
          "instances",
          "rootComponentId"
        ],
        "title": "Manifest",
        "type": "object"
      },
...
      "Project": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "current_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Revision Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "name",
          "current_revision_id",
          "created_at",
          "updated_at"
        ],
        "title": "Project",
        "type": "object"
      },
...
      "Revision": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/Manifest"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "restored_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restored From"
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationReport"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "id",
          "project_id",
          "run_id",
          "ordinal",
          "summary",
          "manifest",
          "created_at",
          "restored_from"
        ],
        "title": "Revision",
        "type": "object"
      },
      "Run": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "base_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision Id"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "waiting_input",
              "paused",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "selected_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Ids",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "model_calls": {
            "title": "Model Calls",
            "type": "integer"
          },
          "backend_version": {
            "title": "Backend Version",
            "type": "integer"
          },
          "execution_environment": {
            "title": "Execution Environment",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "owner_id",
          "base_revision_id",
          "status",
          "message",
          "selected_ids",
          "error",
          "created_at",
          "updated_at",
          "workflow_id",
          "model_calls",
          "backend_version",
          "execution_environment"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunEvent": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "status",
              "tool",
              "validation",
              "error"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Stage"
          },
          "attempt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Attempt"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Elapsed Ms"
          }
        },
        "required": [
          "id",
          "run_id",
          "kind",
          "message",
          "created_at"
        ],
        "title": "RunEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationReport": {
        "properties": {
          "identity": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Identity",
            "type": "object"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementCheck"
            },
            "title": "Requirements",
            "type": "array"
          },
          "allRequirementsVerified": {
            "title": "Allrequirementsverified",
            "type": "boolean"
          }
        },
        "required": [
          "identity",
          "requirements",
          "allRequirementsVerified"
        ],
        "title": "ValidationReport",
        "type": "object"
      },
      "WorkspaceState": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/Revision"
            },
            "title": "Revisions",
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/Run"
            },
            "title": "Runs",
            "type": "array"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "calculations": {
            "items": {
              "$ref": "#/components/schemas/CalculationRecord"
            },
            "title": "Calculations",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "project",
          "revisions",
          "messages",
          "runs",
          "artifacts",
          "calculations",
          "events"
        ],
        "title": "WorkspaceState",
        "type": "object"
      },
...
    }
  }
}